+++ aarch64/isa/add.go
@@ -3,7 +3,7 @@ package aarch64isa
 import (
 	"alon.kr/x/aarch64codegen/instructions"
 	"alon.kr/x/list"
-	aarch64core "alon.kr/x/usm/aarch64/core"
+	aarch64codegen "alon.kr/x/usm/aarch64/codegen"
 	aarch64translation "alon.kr/x/usm/aarch64/translation"
 	"alon.kr/x/usm/core"
 	"alon.kr/x/usm/gen"
@@ -24,7 +24,7 @@ type Add struct {
 	instructions.Add
 }
 
-func (i Add) Generate(*aarch64core.FunctionCodegenContext) instructions.Instruction {
+func (i Add) Generate(*aarch64codegen.FunctionCodegenContext) instructions.Instruction {
 	return i
 }
 
@@ -33,7 +33,7 @@ type AddImm struct {
 	instructions.AddImm
 }
 
-func (i AddImm) Generate(*aarch64core.FunctionCodegenContext) instructions.Instruction {
+func (i AddImm) Generate(*aarch64codegen.FunctionCodegenContext) instructions.Instruction {
 	return i
 }
