+++ aarch64/isa/movz.go
@@ -2,7 +2,7 @@ package aarch64isa
 
 import (
 	"alon.kr/x/aarch64codegen/instructions"
-	aarch64core "alon.kr/x/usm/aarch64/core"
+	aarch64codegen "alon.kr/x/usm/aarch64/codegen"
 	aarch64translation "alon.kr/x/usm/aarch64/translation"
 	"alon.kr/x/usm/core"
 	"alon.kr/x/usm/gen"
@@ -20,7 +20,7 @@ func (Movz) PossibleNextSteps() (gen.StepInfo, core.ResultList) {
 	return gen.StepInfo{PossibleContinue: true}, core.ResultList{}
 }
 
-func (i Movz) Generate(*aarch64core.FunctionCodegenContext) instructions.Instruction {
+func (i Movz) Generate(*aarch64codegen.FunctionCodegenContext) instructions.Instruction {
 	return i
 }
